# docs/openapi.yaml
openapi: 3.1.0
info:
  title: holdfast 예약 API
  version: 1.0.0
  summary: 좌석 단위 점유 제어 예약 흐름의 API 계약
  description: |
    좌석맵 조회 → 좌석 선점(hold) → 예약 확정 → 조회·취소까지의 계약이다.

    이 문서의 목적은 프론트 담당자가 서버 구현을 기다리지 않고 화면을 만들 수 있게
    하는 것이다. 요청·응답 스키마와 오류 코드는 확정이며, 변경은 두 사람의 작업을
    동시에 깨뜨린다. 변경이 필요하면 PR로 이 파일을 먼저 고친다.

    설계 결정의 근거는 `docs/api-spec.md`에 있다.

    ## 확정된 전제

    - **좌석 선택 상태는 클라이언트가 보관한다.** 서버 세션을 쓰지 않는다.
      앱 2대 구성에서 세션 공유는 측정 오염 요인이다 (concurrency-spec 0.4).
    - **모든 상태 변경 요청은 `Idempotency-Key` 헤더를 요구한다** (concurrency-spec 6절).
    - **409는 정상 거절이며 오류가 아니다.** 원인은 `code`로 구분한다.
      경합으로 인한 거절은 어떤 경우에도 5xx로 나가지 않는다 (concurrency-spec 7.1).
    - **모든 시각은 서버 기준이다.** 응답의 `serverTime`을 기준으로 남은 시간을
      계산한다. 클라이언트 시계를 신뢰하지 않는다 (design-spec 5.1).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: 로컬 Docker Compose (nginx → 앱 2대)

# 인증은 요구사항 추적표(design-spec 3.1)에 없어 구현하지 않는다.
# 빈 배열은 "이 API에 인증 요구사항이 없다"는 명시적 선언이다.
# 사용자 식별은 X-User-Id 헤더가 대신하며 그 한계는 파라미터 설명에 적어두었다.
security: []
...
paths:
...
  /api/reservations/{reservationId}/tickets:
    get:
      tags: [tickets]
      operationId: getTickets
      summary: 예약의 티켓 목록
      description: |
        확정된 예약좌석마다 발급된 티켓을 반환한다. 발급은 결제 승인과 같은
        트랜잭션에서 일어나므로(state-transitions.md 4절), 결제가 201로
        응답한 시점에 이미 존재한다.

        본인 예약만 조회할 수 있다. 다른 사용자의 예약이면 존재 여부를 흘리지
        않기 위해 403이 아니라 404를 반환한다(`getReservation`과 같은 규칙).
      parameters:
        - $ref: '#/components/parameters/ReservationIdPath'
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: 티켓 목록. 확정 전이거나 결제 전이면 빈 배열이다.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tickets/scan:
    post:
      tags: [tickets]
      operationId: scanTicket
      summary: 검표
      description: |
        QR 토큰으로 입장을 판정한다. **거절도 200이다** — 중복 사용·시간 외·
        무효 티켓은 오류가 아니라 검표의 정상적인 판정 결과이며, `result`로
        구분한다(concurrency-spec 7.1의 정상 거절·오류 분리 원칙과 같은 이유).

        `Idempotency-Key`를 요구하지 않는다. 검표는 재시도를 멱등하게 눌러야
        하는 요청이 아니라 스캔할 때마다 일어나는 물리적 사건의 기록이다 —
        같은 티켓을 두 번 스캔하면 두 번째는 네트워크 재시도가 아니라 정말
        다시 스캔된 것이고, `REJECTED_DUPLICATE`가 그 사실을 그대로 담는다.

        사용자 식별 헤더도 요구하지 않는다. 스캔하는 주체는 티켓 소유자가
        아니라 게이트 단말이다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketScanRequest'
      responses:
        '200':
          description: 검표 판정. 성공·거절 모두 이 상태로 나간다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketScanResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
...
    ReservationIdPath:
      name: reservationId
      in: path
      required: true
      description: 예약 ID
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 55021
...
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: |
        요청자 식별자. 인증·회원 관리는 요구사항 추적표(design-spec 3.1)에 없어 구현하지
        않으므로 사용자 식별을 이 헤더로 대신한다. 인증이 붙으면 이 헤더는 토큰에서
        유도된 값으로 대체되고 계약의 나머지는 바뀌지 않는다.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 7001
  responses:
    BadRequest:
      description: 요청 형식이 잘못됨
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Bad Request
            status: 400
            detail: seatIds는 비어 있을 수 없습니다.
            code: VALIDATION_FAILED
            serverTime: '2026-09-01T04:16:19Z'
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Not Found
            status: 404
            detail: 회차를 찾을 수 없습니다.
            code: SESSION_NOT_FOUND
            serverTime: '2026-09-01T04:16:19Z'
...
    InternalError:
      description: |
        서버 결함. **경합으로 인한 거절은 절대 여기로 오지 않는다.**
        이 응답이 나오면 측정에서 오류율로 집계된다 (concurrency-spec 7.1).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
            detail: 예기치 못한 오류가 발생했습니다.
            code: INTERNAL_ERROR
            serverTime: '2026-09-01T04:16:19Z'
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        오류·거절 사유 코드. **k6는 이 값으로 태깅해 집계한다.** 문자열이 바뀌면 측정
        스크립트가 함께 깨지므로 값을 변경하거나 다른 의미로 재사용하지 않는다.

        분류는 docs/api-spec.md 3절의 표를 따른다.
      enum:
        - SEAT_ALREADY_SOLD
        - SEAT_HELD_BY_OTHER
        - HOLD_EXPIRED
        - HOLD_RELEASED
        - HOLD_ALREADY_CONFIRMED
        - QUOTA_EXCEEDED
        - RESERVATION_NOT_OPEN
        - RESERVATION_ALREADY_CANCELLED
        - RESERVATION_NOT_CANCELLABLE
        - LOCK_TIMEOUT
        - RETRY_EXHAUSTED
        - IDEMPOTENCY_KEY_REUSED
        - IDEMPOTENCY_KEY_IN_PROGRESS
        - VALIDATION_FAILED
        - SESSION_NOT_FOUND
        - SEAT_NOT_IN_SESSION
        - HOLD_NOT_FOUND
        - RESERVATION_NOT_FOUND
        - INTERNAL_ERROR
    Problem:
      type: object
      description: |
        RFC 9457 Problem Details. Spring Boot 4가 처리하지 못한 예외에도 같은 형식을
        내보내므로 클라이언트가 다루는 오류 표현이 하나로 유지된다.
      required: [status, code]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
          examples: [Conflict]
        status:
          type: integer
          examples: [409]
        detail:
          type: string
          description: 사용자에게 그대로 보여줄 수 있는 한국어 설명.
        instance:
          type: string
          description: 요청 경로
        code:
          $ref: '#/components/schemas/ErrorCode'
        serverTime:
          type: string
          format: date-time
          description: 서버 기준 현재 시각
        conflicts:
          type: array
          description: 좌석별 거절 사유. POST /api/holds의 409에만 들어 있다.
          items:
            $ref: '#/components/schemas/SeatConflict'
    SeatConflict:
      type: object
      required: [seatId, code]
      properties:
        seatId:
          type: integer
          format: int64
          examples: [30412]
        code:
          $ref: '#/components/schemas/ErrorCode'
...
    TicketStatus:
      type: string
      description: |
        티켓 상태 (design-spec 3.3, erd.md ticket.status).

        - ISSUED — 발급됨. 검표를 통과할 수 있는 유일한 상태
        - USED — 검표 통과. 재사용은 U-11이 막는다
        - VOID — 예약 취소로 무효. 이 구현은 lazy 검증을 쓰므로 이 값에
          도달하지 않는다 — 취소된 예약의 티켓은 status가 ISSUED로 남은 채
          검표 시점에 예약 상태를 함께 확인해 거절된다(`TicketScanResult`
          REJECTED_INVALID)
      enum: [ISSUED, USED, VOID]
    Ticket:
      type: object
      required: [ticketId, qrToken, status, seatNo, zoneName, issuedAt, serverTime]
      properties:
        ticketId:
          type: integer
          format: int64
        qrToken:
          type: string
          description: QR에 인코딩할 값. U-10으로 유일성이 보장된다.
        status:
          $ref: '#/components/schemas/TicketStatus'
        seatNo:
          type: string
          examples: [A-12]
        zoneName:
          type: string
          examples: [A구역]
        issuedAt:
          type: string
          format: date-time
        usedAt:
          type: [string, 'null']
          format: date-time
          description: 검표 통과 시각. 미사용이면 null이다.
        serverTime:
          type: string
          format: date-time
    TicketScanRequest:
      type: object
      required: [qrToken]
      properties:
        qrToken:
          type: string
          description: QR에서 읽은 값. Ticket.qrToken과 대조한다.
    ScanResult:
      type: string
      description: |
        검표 판정 (erd.md ticket_scan.result).

        - ADMITTED — 입장 허용. 티켓당 1건만 존재할 수 있다(U-11)
        - REJECTED_DUPLICATE — 이미 사용된 티켓
        - REJECTED_TIME — 회차의 입장 가능 시간이 아님(REQ-06)
        - REJECTED_INVALID — 토큰이 없거나, 예약이 취소됐거나, 티켓이 무효
      enum: [ADMITTED, REJECTED_DUPLICATE, REJECTED_TIME, REJECTED_INVALID]
    TicketScanResult:
      type: object
      description: |
        검표 결과. `REJECTED_*`도 200으로 나간다 — 검표 거절은 오류가 아니라
        판정이다.
      required: [result, scannedAt, serverTime]
      properties:
        result:
          $ref: '#/components/schemas/ScanResult'
        ticketId:
          type: [integer, 'null']
          format: int64
          description: 토큰 자체가 유효하지 않으면 null이다(대조할 티켓이 없다).
        rejectReason:
          type: [string, 'null']
          description: 사용자에게 그대로 보여줄 수 있는 한국어 설명. 성공이면 null이다.
        scannedAt:
          type: string
          format: date-time
        serverTime:
          type: string
          format: date-time
...